+++ src/Geopilot.Frontend/src/api/apiInterfaces.ts
@@ -20,11 +20,6 @@ export class ApiError extends Error {
   }
 }
 
-export interface ApiContextInterface {
-  fetchApi: <T>(url: string, options?: Partial<FetchParams>) => Promise<T>;
-  fetchLocalizedMarkdown: (markdown: string, language: string) => Promise<string | null>;
-}
-
 export enum FieldEvaluationType {
   NotEvaluated = "notEvaluated",
   Optional = "optional",