+++ src/Geopilot.Frontend/src/components/appSettings/appSettingsContext.tsx
@@ -1,8 +1,8 @@
 import { createContext, FC, PropsWithChildren, useEffect, useState } from "react";
 import { AppSettingsContextInterface, ClientSettings } from "./appSettingsInterface";
-import { useApi } from "../../api";
 import { ContentType } from "../../api/apiInterfaces.ts";
 import { useTranslation } from "react-i18next";
+import useFetch from "../../hooks/useFetch.ts";
 
 export const AppSettingsContext = createContext<AppSettingsContextInterface>({
   initialized: false,
@@ -12,17 +12,19 @@ export const AppSettingsContext = createContext<AppSettingsContextInterface>({
 
 export const AppSettingsProvider: FC<PropsWithChildren> = ({ children }) => {
   const { i18n } = useTranslation();
-  const { fetchApi, fetchLocalizedMarkdown } = useApi();
-  const [clientSettings, setClientSettings] = useState<ClientSettings | null | undefined>(undefined);
-  const [termsOfUse, setTermsOfUse] = useState<string | null | undefined>(undefined);
+  const { fetchApi, fetchLocalizedMarkdown } = useFetch();
+  const [clientSettings, setClientSettings] = useState<ClientSettings | null>();
+  const [termsOfUse, setTermsOfUse] = useState<string | null>();
 
   useEffect(() => {
     fetchApi<ClientSettings>("/client-settings.json", { responseType: ContentType.Json })
       .then(setClientSettings)
       .catch(() => setClientSettings(null));
+  }, [fetchApi]);
 
+  useEffect(() => {
     fetchLocalizedMarkdown("terms-of-use", i18n.language).then(setTermsOfUse);
-  }, [fetchApi, fetchLocalizedMarkdown, i18n.language]);
+  }, [fetchLocalizedMarkdown, i18n.language]);
 
   useEffect(() => {
     if (clientSettings) {