+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Modularized a component of our CI/CD Workflows.
 - Updated README.md to work with new dev-cert naming.
 - **BREAKING** Removed references to name "ilicheck", replaced with "INTERLIS".
+- Adjusted our Hooks to comply more with ESLint rules.
 
 ### Fixed